+++ tests/unit/providers/test_configuration_py2_py3.py
@@ -9,6 +9,10 @@ import tempfile
 import unittest2 as unittest
 
 from dependency_injector import containers, providers, errors
+try:
+    import yaml
+except ImportError:
+    yaml = None
 
 
 class ConfigTests(unittest.TestCase):
@@ -395,6 +399,16 @@ class ConfigFromIniTests(unittest.TestCase):
         self.assertEqual(self.config.section2(), {'value2': '2'})
         self.assertEqual(self.config.section2.value2(), '2')
 
+    def test_option(self):
+        self.config.option.from_ini(self.config_file_1)
+
+        self.assertEqual(self.config(), {'option': {'section1': {'value1': '1'}, 'section2': {'value2': '2'}}})
+        self.assertEqual(self.config.option(), {'section1': {'value1': '1'}, 'section2': {'value2': '2'}})
+        self.assertEqual(self.config.option.section1(), {'value1': '1'})
+        self.assertEqual(self.config.option.section1.value1(), '1')
+        self.assertEqual(self.config.option.section2(), {'value2': '2'})
+        self.assertEqual(self.config.option.section2.value2(), '2')
+
     def test_merge(self):
         self.config.from_ini(self.config_file_1)
         self.config.from_ini(self.config_file_2)
@@ -421,6 +435,43 @@ class ConfigFromIniTests(unittest.TestCase):
         self.assertEqual(self.config.section2.value2(), '2')
         self.assertEqual(self.config.section3(), {'value3': '3'})
         self.assertEqual(self.config.section3.value3(), '3')
+
+    def test_file_does_not_exist(self):
+        self.config.from_ini('./does_not_exist.ini')
+        self.assertEqual(self.config(), {})
+
+    def test_file_does_not_exist_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        with self.assertRaises(IOError):
+            self.config.from_ini('./does_not_exist.ini')
+
+    def test_option_file_does_not_exist(self):
+        self.config.option.from_ini('does_not_exist.ini')
+        self.assertIsNone(self.config.option.undefined())
+
+    def test_option_file_does_not_exist_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        with self.assertRaises(IOError):
+            self.config.option.from_ini('./does_not_exist.ini')
+
+    def test_required_file_does_not_exist(self):
+        with self.assertRaises(IOError):
+            self.config.from_ini('./does_not_exist.ini', required=True)
+
+    def test_required_option_file_does_not_exist(self):
+        with self.assertRaises(IOError):
+            self.config.option.from_ini('./does_not_exist.ini', required=True)
+
+    def test_not_required_file_does_not_exist_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        self.config.from_ini('./does_not_exist.ini', required=False)
+        self.assertEqual(self.config(), {})
+
+    def test_not_required_option_file_does_not_exist_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        self.config.option.from_ini('./does_not_exist.ini', required=False)
+        with self.assertRaises(errors.Error):
+            self.config.option()
 
 
 class ConfigFromIniWithEnvInterpolationTests(unittest.TestCase):
@@ -525,6 +576,51 @@ class ConfigFromYamlTests(unittest.TestCase):
         self.assertEqual(self.config.section3(), {'value3': 3})
         self.assertEqual(self.config.section3.value3(), 3)
 
+    @unittest.skipIf(sys.version_info[:2] == (3, 4), 'PyYAML does not support Python 3.4')
+    def test_file_does_not_exist(self):
+        self.config.from_yaml('./does_not_exist.yml')
+        self.assertEqual(self.config(), {})
+
+    @unittest.skipIf(sys.version_info[:2] == (3, 4), 'PyYAML does not support Python 3.4')
+    def test_file_does_not_exist_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        with self.assertRaises(IOError):
+            self.config.from_yaml('./does_not_exist.yml')
+
+    @unittest.skipIf(sys.version_info[:2] == (3, 4), 'PyYAML does not support Python 3.4')
+    def test_option_file_does_not_exist(self):
+        self.config.option.from_yaml('./does_not_exist.yml')
+        self.assertIsNone(self.config.option())
+
+    @unittest.skipIf(sys.version_info[:2] == (3, 4), 'PyYAML does not support Python 3.4')
+    def test_option_file_does_not_exist_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        with self.assertRaises(IOError):
+            self.config.option.from_yaml('./does_not_exist.yml')
+
+    @unittest.skipIf(sys.version_info[:2] == (3, 4), 'PyYAML does not support Python 3.4')
+    def test_required_file_does_not_exist(self):
+        with self.assertRaises(IOError):
+            self.config.from_yaml('./does_not_exist.yml', required=True)
+
+    @unittest.skipIf(sys.version_info[:2] == (3, 4), 'PyYAML does not support Python 3.4')
+    def test_required_option_file_does_not_exist(self):
+        with self.assertRaises(IOError):
+            self.config.option.from_yaml('./does_not_exist.yml', required=True)
+
+    @unittest.skipIf(sys.version_info[:2] == (3, 4), 'PyYAML does not support Python 3.4')
+    def test_not_required_file_does_not_exist_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        self.config.from_yaml('./does_not_exist.yml', required=False)
+        self.assertEqual(self.config(), {})
+
+    @unittest.skipIf(sys.version_info[:2] == (3, 4), 'PyYAML does not support Python 3.4')
+    def test_not_required_option_file_does_not_exist_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        self.config.option.from_yaml('./does_not_exist.yml', required=False)
+        with self.assertRaises(errors.Error):
+            self.config.option()
+
     def test_no_yaml_installed(self):
         @contextlib.contextmanager
         def no_yaml_module():
@@ -580,6 +676,51 @@ class ConfigFromYamlWithEnvInterpolationTests(unittest.TestCase):
         )
         self.assertEqual(self.config.section1(), {'value1': 'test-value'})
         self.assertEqual(self.config.section1.value1(), 'test-value')
+
+    @unittest.skipIf(sys.version_info[:2] == (3, 4), 'PyYAML does not support Python 3.4')
+    def test_option_env_variable_interpolation(self):
+        self.config.option.from_yaml(self.config_file)
+
+        self.assertEqual(
+            self.config.option(),
+            {
+                'section1': {
+                    'value1': 'test-value',
+                },
+            },
+        )
+        self.assertEqual(self.config.option.section1(), {'value1': 'test-value'})
+        self.assertEqual(self.config.option.section1.value1(), 'test-value')
+
+    @unittest.skipIf(sys.version_info[:2] == (3, 4), 'PyYAML does not support Python 3.4')
+    def test_env_variable_interpolation_custom_loader(self):
+        self.config.from_yaml(self.config_file, loader=yaml.UnsafeLoader)
+
+        self.assertEqual(
+            self.config(),
+            {
+                'section1': {
+                    'value1': 'test-value',
+                },
+            },
+        )
+        self.assertEqual(self.config.section1(), {'value1': 'test-value'})
+        self.assertEqual(self.config.section1.value1(), 'test-value')
+
+    @unittest.skipIf(sys.version_info[:2] == (3, 4), 'PyYAML does not support Python 3.4')
+    def test_option_env_variable_interpolation_custom_loader(self):
+        self.config.option.from_yaml(self.config_file, loader=yaml.UnsafeLoader)
+
+        self.assertEqual(
+            self.config.option(),
+            {
+                'section1': {
+                    'value1': 'test-value',
+                },
+            },
+        )
+        self.assertEqual(self.config.option.section1(), {'value1': 'test-value'})
+        self.assertEqual(self.config.option.section1.value1(), 'test-value')
 
 
 class ConfigFromDict(unittest.TestCase):
@@ -641,6 +782,43 @@ class ConfigFromDict(unittest.TestCase):
         self.assertEqual(self.config.section3(), {'value3': '3'})
         self.assertEqual(self.config.section3.value3(), '3')
 
+    def test_empty_dict(self):
+        self.config.from_dict({})
+        self.assertEqual(self.config(), {})
+
+    def test_option_empty_dict(self):
+        self.config.option.from_dict({})
+        self.assertEqual(self.config.option(), {})
+
+    def test_empty_dict_in_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        with self.assertRaises(ValueError):
+            self.config.from_dict({})
+
+    def test_option_empty_dict_in_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        with self.assertRaises(ValueError):
+            self.config.option.from_dict({})
+
+    def test_required_empty_dict(self):
+        with self.assertRaises(ValueError):
+            self.config.from_dict({}, required=True)
+
+    def test_required_option_empty_dict(self):
+        with self.assertRaises(ValueError):
+            self.config.option.from_dict({}, required=True)
+
+    def test_not_required_empty_dict_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        self.config.from_dict({}, required=False)
+        self.assertEqual(self.config(), {})
+
+    def test_not_required_option_empty_dict_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        self.config.option.from_dict({}, required=False)
+        self.assertEqual(self.config.option(), {})
+        self.assertEqual(self.config(), {'option': {}})
+
 
 class ConfigFromEnvTests(unittest.TestCase):
 
@@ -656,13 +834,77 @@ class ConfigFromEnvTests(unittest.TestCase):
         self.config.from_env('CONFIG_TEST_ENV')
         self.assertEqual(self.config(), 'test-value')
 
-    def test_default(self):
-        self.config.from_env('UNDEFINED_ENV', 'default-value')
-        self.assertEqual(self.config(), 'default-value')
-
     def test_with_children(self):
         self.config.section1.value1.from_env('CONFIG_TEST_ENV')
 
         self.assertEqual(self.config(), {'section1': {'value1': 'test-value'}})
         self.assertEqual(self.config.section1(), {'value1': 'test-value'})
         self.assertEqual(self.config.section1.value1(), 'test-value')
+
+    def test_default(self):
+        self.config.from_env('UNDEFINED_ENV', 'default-value')
+        self.assertEqual(self.config(), 'default-value')
+
+    def test_default_none(self):
+        self.config.from_env('UNDEFINED_ENV')
+        self.assertIsNone(self.config())
+
+    def test_option_default_none(self):
+        self.config.option.from_env('UNDEFINED_ENV')
+        self.assertIsNone(self.config.option())
+
+    def test_undefined_in_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        with self.assertRaises(ValueError):
+            self.config.from_env('UNDEFINED_ENV')
+
+    def test_option_undefined_in_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        with self.assertRaises(ValueError):
+            self.config.option.from_env('UNDEFINED_ENV')
+
+    def test_undefined_in_strict_mode_with_default(self):
+        self.config = providers.Configuration(strict=True)
+        self.config.from_env('UNDEFINED_ENV', 'default-value')
+        self.assertEqual(self.config(), 'default-value')
+
+    def test_option_undefined_in_strict_mode_with_default(self):
+        self.config = providers.Configuration(strict=True)
+        self.config.option.from_env('UNDEFINED_ENV', 'default-value')
+        self.assertEqual(self.config.option(), 'default-value')
+
+    def test_required_undefined(self):
+        with self.assertRaises(ValueError):
+            self.config.from_env('UNDEFINED_ENV', required=True)
+
+    def test_required_undefined_with_default(self):
+        self.config.from_env('UNDEFINED_ENV', default='default-value', required=True)
+        self.assertEqual(self.config(), 'default-value')
+
+    def test_option_required_undefined(self):
+        with self.assertRaises(ValueError):
+            self.config.option.from_env('UNDEFINED_ENV', required=True)
+
+    def test_option_required_undefined_with_default(self):
+        self.config.option.from_env('UNDEFINED_ENV', default='default-value', required=True)
+        self.assertEqual(self.config.option(), 'default-value')
+
+    def test_not_required_undefined_in_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        self.config.from_env('UNDEFINED_ENV', required=False)
+        self.assertIsNone(self.config())
+
+    def test_option_not_required_undefined_in_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        self.config.option.from_env('UNDEFINED_ENV', required=False)
+        self.assertIsNone(self.config.option())
+
+    def test_not_required_undefined_with_default_in_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        self.config.from_env('UNDEFINED_ENV', default='default-value', required=False)
+        self.assertEqual(self.config(), 'default-value')
+
+    def test_option_not_required_undefined_with_default_in_strict_mode(self):
+        self.config = providers.Configuration(strict=True)
+        self.config.option.from_env('UNDEFINED_ENV', default='default-value', required=False)
+        self.assertEqual(self.config.option(), 'default-value')